% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ability_scores.R
\name{ability_singular}
\alias{ability_singular}
\title{Rolls for a Single Ability Score}
\usage{
ability_singular(method = "4d6")
}
\arguments{
\item{method}{(character) string of "4d6", "3d6", or "1d20" ("d20" also accepted). Enter your preferred method of rolling for each ability score ("4d6" drops lowest before summing)}
}
\value{
(numeric) vector of roll outcomes (not summed)
}
\description{
Rolls for a single ability score using the specified method of dice rolling.
}
