% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_block.R
\name{accepted_classes}
\alias{accepted_classes}
\title{Return Vector of Accepted Classes}
\usage{
accepted_classes()
}
\value{
character vector of accepted class names
}
\description{
Simply returns a vector of classes that `class_block()` accepts currently. Submit an issue on the GitHub repository (github.com/njlyon0/dndR) if you want a class added.
}
\examples{
# Want to check which classes this package supports?
accepted_classes()

}
