% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_convert.R
\name{cr_convert}
\alias{cr_convert}
\title{Convert Challenge Rating to Experience Points}
\usage{
cr_convert(cr = NULL)
}
\arguments{
\item{cr}{(numeric) Challenge rating that you want to calculate experience points for}
}
\description{
Converts challenge rating (CR) into experience points (XP) using two formulas for a parabola (one for less than/equal to 20 and one for greater than 20). The relationship between CR and XP in the Dungeon Master's Guide (DMG) is disjointed in this way so this is a reasonable move.
}
