% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_functions.R
\name{dnd_classes}
\alias{dnd_classes}
\title{Return Vector of Accepted Classes}
\usage{
dnd_classes()
}
\value{
(character) vector of accepted class names
}
\description{
Simply returns a vector of classes that `class_block()` accepts currently. Submit an issue on the GitHub repository if you want a class added.
}
\examples{
# Want to check which classes this package supports?
dndR::dnd_classes()

}
