% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage_types.R
\name{dnd_damage_types}
\alias{dnd_damage_types}
\title{Return Vector of Supported DnD Damage Types}
\usage{
dnd_damage_types()
}
\value{
character vector of damage types
}
\description{
Simply returns a vector of damage types in DnD
}
