% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_functions.R
\name{dnd_races}
\alias{dnd_races}
\title{Return Vector of Supported DnD Races}
\usage{
dnd_races()
}
\value{
(character) vector of supported race designations
}
\description{
Simply returns a vector of races that `race_mods()` accepts currently. Submit an issue on the GitHub repository if you want a race added.
}
\examples{
# Want to check which races this package supports?
dndR::dnd_races()

}
