% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_functions.R
\name{first_names_fem}
\alias{first_names_fem}
\title{Return Vector of Feminine First Names}
\usage{
first_names_fem()
}
\value{
(character) vector of first names
}
\description{
Simply returns a vector of feminine first names (in alphabetical order)
}
\examples{
# Full set of feminine first names
dndR::first_names_fem()

}
