% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_functions.R
\name{first_names_masc}
\alias{first_names_masc}
\title{Return Vector of Masculine First Names}
\usage{
first_names_masc()
}
\value{
(character) vector of first names
}
\description{
Simply returns a vector of masculine first names (in alphabetical order)
}
\examples{
# Full set of masculine first names
dndR::first_names_masc()

}
