% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monster_stats.R
\name{monster_stats}
\alias{monster_stats}
\title{Quickly Identify Monster Statistics}
\usage{
monster_stats(xp = NULL, cr = NULL)
}
\arguments{
\item{xp}{(numeric) Experience Point value of the monster}

\item{cr}{(numeric) Challenge Rating of the monster. Note that this is NOT necessary if XP is provided}
}
\description{
Quickly identify the vital statistics of a single creature worth the provided experience points (XP) or Challenge Rating (CR). Uses the table provided in p. 274-275 of the Dungeon Master's Guide. Accepts Challenge Ratings of 0, '1/8', '1/4, and '1/2' in addition to numbers between 1 and 30. CR is *not necessary* to provide **if** XP is provided.
}
