% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_creator.R
\name{npc_creator}
\alias{npc_creator}
\title{Create a Non-Player Character (NPC)}
\usage{
npc_creator(npc_count = 1)
}
\arguments{
\item{npc_count}{(numeric) number of NPCs for which to choose race/positions}
}
\value{
(dataframe) dataframe with two columns (one for race and one for job) and a number of rows equal to `npc_count`
}
\description{
Randomly selects a race and job for a user-specified number of NPCs
}
\examples{
# Create some information for an NPC
dndR::npc_creator(npc_count = 1)

}
