% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/party_diagram.R
\name{party_diagram}
\alias{party_diagram}
\title{Generate a Diagram of a Party's Ability Scores}
\usage{
party_diagram(pc_stats, by)
}
\arguments{
\item{by}{(character) Either "player" (default) or "ability". Define the facets of the party diagram.}
}
\value{
(ggplot object) The party diagram as a ggplot object.
}
\description{
Input a party's ability scores and visualize either by ability or player character.
}
