% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_creator.R
\name{pc_creator}
\alias{pc_creator}
\title{Create a Player Character (PC)}
\usage{
pc_creator(
  class = NULL,
  race = NULL,
  score_method = "4d6",
  scores_rolled = TRUE,
  scores_df = NULL
)
}
\arguments{
\item{class}{character string of class (supported classes returned by `accepted_classes()`)}

\item{race}{character string of race (supported classes returned by `dnd_races()`)}

\item{score_method}{character string of "4d6", "3d6", or "1d20" ("d20" also accepted). Only values accepted by `ability_scores()` are accepted here}

\item{scores_rolled}{logical indicating whether ability scores have previously been rolled (via `ability_scores()`). Defaults to FALSE}

\item{scores_df}{if 'scores_rolled' is TRUE, the name of the dataframe returned by `ability_scores()`}
}
\value{
a dataframe of five columns and six rows
}
\description{
Stat out a player character of specified race and class using your preferred method for rolling ability scores.
}
\examples{
# Create a PC's base statistics from scratch
pc_creator(class = 'barbarian', race = 'half orc', score_method = "4d6")

# Or you can roll separately and then create a character with that dataframe
my_scores <- ability_scores(method = "4d6")
pc_creator(class = 'sorcerer', race = 'dragonborn', scores_rolled = TRUE, scores_df = my_scores)

}
