% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_plot.R
\name{probability_plot}
\alias{probability_plot}
\title{Generate a Plot of the Frequency of Roll Outcomes}
\usage{
probability_plot(dice = "2d20", roll_num = 999)
}
\arguments{
\item{dice}{(character) specifying the number of dice and which type (e.g., "2d4" for two, four-sided dice). Defaults to two, six-sided dice}

\item{roll_num}{(integer) number of times to roll the specified dice to generate the data fro the probability plot. Defaults to 999}
}
\value{
(ggplot object) roll outcome frequency as a ggplot2 object
}
\description{
Input the number and type of dice to roll and the number of times to roll the dice. This is used to generate a plot of the real distribution of dice outcomes and create a ggplot2 plot of that result. A vertical dashed line is included at the median roll result. Note that low numbers of rolls may not generate realistic frequencies of outcomes
}
\examples{
# Generate a probability plot of 3d8
dndR::probability_plot(dice = "3d8", roll_num = 99)

}
