% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_mods.R
\name{race_mods}
\alias{race_mods}
\title{Identify Race-Based Ability Modifiers}
\usage{
race_mods(race = NULL)
}
\arguments{
\item{race}{(character) string of race (supported classes returned by `dndR::dnd_races()`). Also supports "random" and will randomly select a supported race}
}
\value{
(dataframe) two columns and as many rows as there are abilities modified by the race
}
\description{
Identify the race-based ability modifiers identified in the Player's Handbook (PHB).
}
\examples{
# Identifies race modifiers of provided race
dndR::race_mods(race = "mountain dwarf")

}
