% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reroll.R
\name{reroll}
\alias{reroll}
\title{Re-Roll 1s from a Prior Dice Roll}
\usage{
reroll(dice_faces, first_result = NULL)
}
\arguments{
\item{dice_faces}{(numeric) number of faces on the die/dice to re-roll}

\item{first_result}{(numeric) vector of original dice results (including 1s to reroll)}
}
\value{
(numeric) vector of non-1 original dice results with re-rolled dice results appended
}
\description{
Re-rolls only the dice that "landed on" 1 from a prior use of `roll`. Retains other dice results from the first roll but replaces the ones.
}
