% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll Any Number of Dice}
\usage{
roll(dice = "d20", show_dice = FALSE)
}
\arguments{
\item{dice}{(character) specifying the number of dice and which type (e.g., "2d4" for two, four-sided dice). Defaults to a single twenty-sided die}

\item{show_dice}{(logical) whether to print the values of each individual die included in the total. Defaults to FALSE}
}
\value{
(numeric) sum of specified dice outcomes
}
\description{
Rolls the Specified Number and Type of Dice
}
\examples{
# Roll your desired dice (i.e., randomly sample the specified die)
roll(dice = "4d6")

# Returned as a number so you can add rolls together or integers
roll('1d20') + 5
roll('2d8') + roll('1d4')

}
