% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_block.R
\name{stat_block}
\alias{stat_block}
\title{Assign Ability Scores Based on Class}
\usage{
stat_block(class = NULL, scores_rolled = FALSE, scores_df = NULL)
}
\arguments{
\item{class}{character string of class (supported classes returned by `accepted_classes()`)}

\item{scores_rolled}{logical indicating whether ability scores have previously been rolled (via `ability_scores()`). Defaults to FALSE.}

\item{scores_df}{if 'scores_rolled' is TRUE, the name of the dataframe returned by `ability_scores()`}
}
\value{
a dataframe of two columns and six rows
}
\description{
Assign rolled ability scores based on the recommendations for quick class building given in the Player's Handbook (PHB).
}
