% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_functions.R
\name{surnames}
\alias{surnames}
\title{Return Vector of Surnames}
\usage{
surnames()
}
\value{
(character) vector of last names
}
\description{
Simply returns a vector of surnames (a.k.a. last names) in alphabetical order. Names contributed by William Mitchell.
}
\examples{
# Full set of feminine first names
dndR::surnames()

}
