% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ability_score.R
\name{ability_scores}
\alias{ability_scores}
\title{Roll for all Ability Scores}
\usage{
ability_scores(method = "4d6")
}
\arguments{
\item{method}{a character string of "4d6", "3d6", or "1d20" ("d20" also accepted). Enter your preferred method of rolling for each ability score ("4d6" drops lowest before summing)}
}
\value{
a dataframe of two columns and six rows
}
\description{
Rolls 4d6 and drops the lowest value for six abilities. Doesn't assign abilities to facilitate player selection of which score should be each ability for a given character. Prints a warning if the total of all abilities is less than 70 or if any one ability is less than 8.
}
\examples{
# Roll ability scores using four d6 and dropping the lowest
ability_scores(method = "4d6")

# Roll using 3d6 and dropping nothing
ability_scores("3d6")

# Or if you're truly wild, just roll a d20 for each ability
ability_scores('d20')

}
