% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndR-package.R
\docType{package}
\name{dndR-package}
\alias{dndR}
\alias{dndR-package}
\title{dndR: Dungeons & Dragons Functions for Players and Dungeon Masters}
\description{
The goal of 'dndR' is to provide a suite of Dungeons & Dragons related functions. This package is meant to be useful both to players and Dungeon Masters (DMs). All functions currently focus on Fifth Edition (a.k.a. "5e") but once the next edition is published functions will likely be expanded to include any rule changes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://njlyon0.github.io/dndR-website/}
  \item \url{https://github.com/njlyon0/dndR}
  \item Report bugs at \url{https://github.com/njlyon0/dndR/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Lyon \email{njlyon@alumni.iastate.edu} (\href{https://orcid.org/0000-0003-3905-1078}{ORCID}) (https://njlyon0.github.io/) [copyright holder]

Other contributors:
\itemize{
  \item Tim Schatto-Eckrodt (https://kudusch.de/) [contributor]
  \item Humberto Nappo (\href{https://orcid.org/0000-0001-7810-1635}{ORCID}) [contributor]
}

}
\keyword{internal}
