% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dndR-package.R
\docType{package}
\name{dndR-package}
\alias{dndR}
\alias{dndR-package}
\title{dndR: Dungeons & Dragons Functions for Players and Dungeon Masters}
\description{
The goal of 'dndR' is to provide a suite of Dungeons & Dragons related functions. This package is meant to be useful both to players and Dungeon Masters (DMs). Some functions apply to many tabletop role-playing games (e.g., dice rolling), but others are focused on Fifth Edition (a.k.a. "5e") and where possible both the 2014 and 2024 versions are supported.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://njlyon0.github.io/dndR/}
  \item Report bugs at \url{https://github.com/njlyon0/dndR/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Lyon \email{njlyon@alumni.iastate.edu} (\href{https://orcid.org/0000-0003-3905-1078}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Tim Schatto-Eckrodt (https://kudusch.de)
}

Other contributors:
\itemize{
  \item Humberto Nappo (\href{https://orcid.org/0000-0001-7810-1635}{ORCID}) [contributor]
}

}
\keyword{internal}
