% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npc_creator.R
\name{npc_creator}
\alias{npc_creator}
\title{Create a Non-Player Character (NPC)}
\usage{
npc_creator(npc_count = 1)
}
\arguments{
\item{npc_count}{(numeric) number of NPCs for which to return names, species, and positions}
}
\value{
(dataframe) dataframe with three columns (one each for name, race, and job) and a number of rows equal to `npc_count`
}
\description{
Randomly generates a name for a user-specified number of Non-Player Characters (NPCs) as well as a species (a.k.a. "race" in 2014 rules) and a job
}
\examples{
# Create some information for an NPC
dndR::npc_creator(npc_count = 1)

}
