% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{Roll Any Number of Dice}
\usage{
roll(what = "d20")
}
\arguments{
\item{what}{a character specifying the number of dice and which type (e.g., "2d4" for two, four-sided dice). Defaults to a twenty-sided dice}
}
\value{
Numeric, the sum of specified dice outcomes
}
\description{
Rolls the Specified Number and Type of Dice
}
