% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_block.R
\name{stat_block}
\alias{stat_block}
\title{Create Full Stat Block}
\usage{
stat_block()
}
\value{
a dataframe of two columns and six rows
}
\description{
Rolls 4d6 and drops the lowest value for six abilities. Doesn't assign abilities to facilitate player selection of which score should be each ability for a given character. Prints a warning if the total of all abilities is less than 70 or if any one ability is less than 8.
}
